fixed path joiValidation and default value
