+++ models/contact.js
@@ -1,5 +1,4 @@
 const { Schema, model } = require("mongoose");
-const Joi = require("joi");
 
 const phoneRegexp = /^(\()?\d{3}(\))?(-|\s)?\d{3}(-|\s)\d{4}$/;
 const contactSchema = new Schema(
@@ -28,18 +27,6 @@ const contactSchema = new Schema(
   { versionKey: false, timestamps: true }
 );
 
-const joiSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().email().required(),
-  phone: Joi.string().pattern(phoneRegexp),
-
-  favorite: Joi.bool(),
-});
-
-const updateFavorite = Joi.object({
-  favorite: Joi.bool().required(),
-});
-
 const Contact = model("contact", contactSchema);
 
-module.exports = { Contact, joiSchema, updateFavorite };
+module.exports = { Contact };
