+++ routes/api/auth.js
@@ -2,7 +2,7 @@ const express = require("express");
 
 const { auth, validation, ctrlWrapper } = require("../../middlewares");
 const { auth: ctrl } = require("../../controllers");
-const { joiRegisterSchema, joiLoginSchema } = require("../../models/user");
+const { joiRegisterSchema, joiLoginSchema } = require("../../utils/validation");
 
 const router = express.Router();
