+++ utils/validation/index.js
@@ -1,0 +1,12 @@
+const {
+  joiRegisterSchema,
+  joiLoginSchema,
+} = require("./authValidationSchemas");
+const { joiSchema, updateFavorite } = require("./contactValidationSchema");
+
+module.exports = {
+  joiRegisterSchema,
+  joiLoginSchema,
+  joiSchema,
+  updateFavorite,
+};
